+++ src/internal.hpp
@@ -70,6 +70,7 @@ extern "C" {
 #include "instantiate.hpp"
 #include "internal.hpp"
 #include "level.hpp"
+#include "lidruptracer.hpp"
 #include "limit.hpp"
 #include "logging.hpp"
 #include "lratbuilder.hpp"
@@ -215,6 +216,7 @@ struct Internal {
   vector<Watches> wtab;         // table of watches for all literals
   Clause *conflict;             // set in 'propagation', reset in 'analyze'
   Clause *ignore;               // ignored during 'vivify_propagate'
+  Clause *dummy_binary;         // Dummy binary clause for subsumption
   Clause *external_reason;      // used as reason at external propagations
   Clause *newest_clause;        // used in external_propagate
   bool force_no_backtrack;      // for new clauses with external propagator
@@ -617,13 +619,12 @@ struct Internal {
   //
   bool minimize_literal (int lit, int depth = 0);
   void minimize_clause ();
-  void calculate_minimize_chain (int lit);
+  void calculate_minimize_chain (int lit, std::vector<int> &stack);
 
   // Learning from conflicts in 'analyze.cc'.
   //
   void learn_empty_clause ();
   void learn_unit_clause (int lit);
-  void learn_external_propagated_unit_clause (int lit);
 
   void bump_variable (int lit);
   void bump_variables ();
@@ -653,8 +654,6 @@ struct Internal {
 
   // Learning from external propagator in 'external_propagate.cpp'
   //
-  void elevate_lit_external (int, Clause *);
-  void elevate_original_unit (uint64_t, int);
   bool external_propagate ();
   bool external_check_solution ();
   void add_external_clause (int propagated_lit = 0,
@@ -808,7 +807,8 @@ struct Internal {
   void flush_vivification_schedule (Vivifier &);
   bool consider_to_vivify_clause (Clause *candidate, bool redundant_mode);
   void vivify_analyze_redundant (Vivifier &, Clause *start, bool &);
-  void vivify_build_lrat (int, Clause *);
+  void vivify_build_lrat (int, Clause *,
+                          std::vector<std::tuple<int, Clause *, bool>> &);
   void vivify_chain_for_units (int lit, Clause *reason);
   bool vivify_all_decisions (Clause *candidate, int subsume);
   void vivify_post_process_analysis (Clause *candidate, int subsume);
@@ -1357,8 +1357,8 @@ struct Internal {
 
   double solve_time (); // accumulated time spent in 'solve ()'
 
-  double process_time (); // since solver was initialized
-  double real_time ();    // since solver was initialized
+  double process_time () const; // since solver was initialized
+  double real_time () const;    // since solver was initialized
 
   double time () { return opts.realtime ? real_time () : process_time (); }