+++ src/shrink.cpp
@@ -445,6 +445,7 @@ void Internal::shrink_and_minimize_clause () {
 #if defined(LOGGING) || !defined(NDEBUG)
   const unsigned old_size = clause.size ();
 #endif
+  std::vector<int> stack;
   {
     std::vector<int>::size_type i = 1;
     for (std::vector<int>::size_type j = 1; j < clause.size (); ++j) {
@@ -454,7 +455,7 @@ void Internal::shrink_and_minimize_clause () {
         assert (j < old_clause_lrat.size ());
         assert (mini_chain.empty ());
         if (clause[j] != old_clause_lrat[j]) {
-          calculate_minimize_chain (-old_clause_lrat[j]);
+          calculate_minimize_chain (-old_clause_lrat[j], stack);
           for (auto p : mini_chain) {
             minimize_chain.push_back (p);
           }