+++ src/solver.cpp
@@ -648,6 +648,7 @@ void Solver::reset_assumptions () {
   REQUIRE_VALID_STATE ();
   transition_to_steady_state ();
   external->reset_assumptions ();
+  external->reset_concluded ();
   LOG_API_CALL_END ("reset_assumptions");
 }
 
@@ -656,6 +657,7 @@ void Solver::reset_constraint () {
   REQUIRE_VALID_STATE ();
   transition_to_steady_state ();
   external->reset_constraint ();
+  external->reset_concluded ();
   LOG_API_CALL_END ("reset_constraint");
 }
 
@@ -699,6 +701,7 @@ int Solver::call_external_solve_and_check_results (bool preprocess_only) {
   if (!res) {
     external->reset_assumptions ();
     external->reset_constraint ();
+    external->reset_concluded ();
   }
   return res;
 }