+++ src/vivify.cpp
@@ -631,6 +631,8 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
   if (c->garbage)
     return;
 
+  auto &lrat_stack = vivifier.lrat_stack;
+
   // First check whether the candidate clause is already satisfied and at
   // the same time copy its non fixed literals to 'sorted'.  The literals
   // in the candidate clause might not be sorted anymore due to replacing
@@ -829,7 +831,7 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
           clear_analyzed_literals ();
           if (lrat) {
             assert (lrat_chain.empty ());
-            vivify_build_lrat (lit, v.reason);
+            vivify_build_lrat (lit, v.reason, lrat_stack);
             clear_analyzed_literals ();
           }
 
@@ -877,7 +879,7 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
         clear_analyzed_literals ();
         if (lrat) {
           assert (lrat_chain.empty ());
-          vivify_build_lrat (0, conflict);
+          vivify_build_lrat (0, conflict, lrat_stack);
           clear_analyzed_literals ();
         }
       }
@@ -909,8 +911,8 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
         // strengthen clause
         if (lrat) {
           assert (lrat_chain.empty ());
-          vivify_build_lrat (0, c);
-          vivify_build_lrat (0, conflict);
+          vivify_build_lrat (0, c, lrat_stack);
+          vivify_build_lrat (0, conflict, lrat_stack);
           clear_analyzed_literals ();
         }
         remove = lit;
@@ -1006,7 +1008,7 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
 
     if (lrat && lrat_chain.empty ()) {
       assert (lrat_chain.empty ());
-      vivify_build_lrat (0, c);
+      vivify_build_lrat (0, c, lrat_stack);
       clear_analyzed_literals ();
     }
     vivify_strengthen (c);
@@ -1029,31 +1031,63 @@ void Internal::vivify_clause (Vivifier &vivifier, Clause *c) {
 // the function with (0, c). originally we justified each literal in c on
 // its own but this is not actually necessary.
 //
-void Internal::vivify_build_lrat (int lit, Clause *reason) {
-  LOG (reason, "VIVIFY LRAT justifying %d with reason", lit);
-
-  for (const auto &other : *reason) {
-    if (other == lit)
-      continue;
-    Var &v = var (other);
-    Flags &f = flags (other);
-    if (f.seen)
-      continue;                 // we would like to assert this:
-    analyzed.push_back (other); // assert (val (other) < 0);
-    f.seen = true;              // but we cannot because we have
-    if (!v.level) {             // already backtracked (sometimes)
-      const unsigned uidx =
-          vlit (-other);                // nevertheless we can use var (l)
-      uint64_t id = unit_clauses[uidx]; // as if l was still assigned
-      assert (id);                      // because var is updated lazily
-      lrat_chain.push_back (id);
+// Non-recursive version, as some bugs have been found.  DFS over the
+// reasons with preordering (aka we explore the entire reason before
+// exploring deeper)
+void Internal::vivify_build_lrat (
+    int lit, Clause *reason,
+    std::vector<std::tuple<int, Clause *, bool>> &stack) {
+  assert (stack.empty ());
+  stack.push_back ({lit, reason, false});
+  while (!stack.empty ()) {
+    int lit;
+    Clause *reason;
+    bool finished;
+    std::tie (lit, reason, finished) = stack.back ();
+    LOG ("VIVIFY LRAT justifying %d", lit);
+    stack.pop_back ();
+    if (lit && flags (lit).seen) {
+      LOG ("skipping already justified");
       continue;
     }
-    if (v.reason) { // recursive justification
-      vivify_build_lrat (other, v.reason);
+    if (finished) {
+      lrat_chain.push_back (reason->id);
+      if (lit && reason) {
+        Flags &f = flags (lit);
+        f.seen = true;
+        analyzed.push_back (lit); // assert (val (other) < 0);
+        assert (flags (lit).seen);
+      }
+      continue;
+    } else
+      stack.push_back ({lit, reason, true});
+    for (const auto &other : *reason) {
+      if (other == lit)
+        continue;
+      Var &v = var (other);
+      Flags &f = flags (other);
+      if (f.seen)
+        continue; // we would lik // assert (val (other) < 0);e to assert
+                  // this:
+      // analyzed.push_back (other); // assert (val (other) < 0);
+      // f.seen = true;              // but we cannot because we have
+      if (!v.level) { // already backtracked (sometimes)
+        const unsigned uidx =
+            vlit (-other);                // nevertheless we can use var (l)
+        uint64_t id = unit_clauses[uidx]; // as if l was still assigned
+        assert (id);                      // because var is updated lazily
+        lrat_chain.push_back (id);
+        f.seen = true;
+        analyzed.push_back (other);
+        continue;
+      }
+      if (v.reason) { // recursive justification
+        LOG ("VIVIFY LRAT pushing %d", other);
+        stack.push_back ({other, v.reason, false});
+      }
     }
   }
-  lrat_chain.push_back (reason->id);
+  stack.clear ();
 }
 
 // calculate lrat_chain